+++ libs/model/src/comment/index.ts
@@ -2,3 +2,4 @@ export * from './CreateComment.command';
 export * from './CreateCommentReaction.command';
 export * from './GetComments.query';
 export * from './SearchComments.query';
+export * from './UpdateComment.command';
