+++ libs/schemas/src/commands/comment.schemas.ts
@@ -17,6 +17,15 @@ export const CreateComment = {
   output: Comment.extend({ community_id: z.string() }),
 };
 
+export const UpdateComment = {
+  input: z.object({
+    comment_id: PG_INT,
+    text: z.string().trim().min(1),
+    discord_meta: DiscordMetaSchema.optional(),
+  }),
+  output: Comment.extend({ community_id: z.string() }),
+};
+
 export const CommentCanvasReaction = z.object({
   comment_id: PG_INT,
   reaction: z.enum(['like']),
