+++ packages/commonwealth/server/controllers/server_comments_controller.ts
@@ -9,11 +9,6 @@ import {
   SearchCommentsResult,
   __searchComments,
 } from './server_comments_methods/search_comments';
-import {
-  UpdateCommentOptions,
-  UpdateCommentResult,
-  __updateComment,
-} from './server_comments_methods/update_comment';
 
 /**
  * A controller class containing methods relating to comments
@@ -33,16 +28,6 @@ export class ServerCommentsController {
     options: SearchCommentsOptions,
   ): Promise<SearchCommentsResult> {
     return __searchComments.call(this, options);
-  }
-
-  /**
-   * Updates a comment.
-   *
-   */
-  async updateComment(
-    options: UpdateCommentOptions,
-  ): Promise<UpdateCommentResult> {
-    return __updateComment.call(this, options);
   }
 
   /**
